NODE-1349: Move `EraObservedBehavior` and `ObservedValidatorBehavior` to `models` submodule.
