+++ casper/src/main/scala/io/casperlabs/casper/dag/DagOperations.scala
@@ -5,13 +5,12 @@ import cats.implicits._
 import cats.{Eq, Eval, Monad}
 import com.google.protobuf.ByteString
 import io.casperlabs.casper.Estimator.BlockHash
-import io.casperlabs.casper.PrettyPrinter
 import io.casperlabs.casper.consensus.{Block, BlockSummary, Era}
 import io.casperlabs.casper.dag.EraObservedBehavior.{LocalDagView, MessageJPast}
 import io.casperlabs.casper.highway.MessageProducer
 import io.casperlabs.casper.util.ProtoUtil
 import io.casperlabs.catscontrib.MonadThrowable
-import io.casperlabs.models.Message
+import io.casperlabs.models.{Message, ObservedValidatorBehavior}
 import io.casperlabs.models.Message.asJRank
 import io.casperlabs.shared.Sorting.jRankOrdering
 import io.casperlabs.shared.StreamT
@@ -26,6 +25,8 @@ import scala.collection.mutable
 import io.casperlabs.storage.dag.AncestorsStorage
 import io.casperlabs.storage.dag.AncestorsStorage.Relation
 import io.casperlabs.shared.ByteStringPrettyPrinter._
+
+import scala.util.control.NoStackTrace
 
 object DagOperations {
 
@@ -519,7 +520,141 @@ object DagOperations {
     message.justifications.toList
       .map(_.latestBlockHash)
       .traverse(dag.lookupUnsafe(_))
-      .flatMap(EraObservedBehavior.messageJPast[F](dag, _, erasObservedBehavior))
+      .flatMap(messageJPast[F](dag, _, erasObservedBehavior))
+
+  /**
+    * Calculates panorama of a set of justifications.
+    *
+    * Panorama is a "view" into the past of the message (j-past-cone).
+    * It is the latest message (or multiple messages) per validator seen
+    * in the j-past-cone of the message.
+    *
+    * We start with the `eraObservedBehavior` which is a local view of the DAG.
+    * It contains superset of what the `justificaions` may point at.
+    *
+    * @param dag
+    * @param justifications
+    * @param erasObservedBehavior
+    * @return
+    */
+  def messageJPast[F[_]: MonadThrowable](
+      dag: DagLookup[F],
+      justifications: List[Message],
+      erasObservedBehavior: LocalDagView[Message]
+  ): F[MessageJPast[Message]] = {
+
+    type EraId = ByteString
+
+    import ObservedValidatorBehavior._
+    import EraObservedBehavior._
+
+    // Map a message's direct justifications to a map that represents its j-past-cone view.
+    val toEraMap: List[Message] => Map[EraId, Map[ByteString, Set[Message]]] =
+      _.groupBy(_.eraId).mapValues(_.groupBy(_.validatorId).mapValues(_.toSet))
+
+    def empty(v: ByteString): (ByteString, Set[Message]) =
+      (v, Set.empty[Message])
+    def honest(v: ByteString, m: Message): (ByteString, Set[Message]) =
+      (v, Set(m))
+    def equivocated(v: ByteString, msgs: Set[Message]): (ByteString, Set[Message]) =
+      (v, msgs)
+
+    // A map from every observed era to a set of validators that produced a message in each of the eras.
+    // We will use it later when merging two p-cones and not skip validators not visible in the direct justifications.
+    val baseMap =
+      erasObservedBehavior.erasValidators
+        .filterNot(_._1.isEmpty) // filter out Genesis
+        .mapValues(_.toList.map(_ -> Set.empty[Message]).toMap)
+
+    val observedKeyBlocks =
+      erasObservedBehavior.keyBlockHashes.map(_.show).mkString("[", ", ", "]")
+
+    (baseMap |+| toEraMap(justifications.filterNot(_.isGenesisLike))).toList
+      .traverse {
+        case (era, validatorsLatestMessages) =>
+          validatorsLatestMessages.toList
+            .traverse {
+              case (validator, messages) =>
+                erasObservedBehavior.getStatus(era, validator) match {
+                  case None =>
+                    val msg = s"Message directly cites validator " +
+                      s"${validator.show} in an era ${era.show} " +
+                      s"but expected messages only from $observedKeyBlocks eras."
+                    MonadThrowable[F].raiseError[(ByteString, Set[Message])](
+                      new IllegalStateException(msg) with NoStackTrace
+                    )
+                  case Some(Empty) =>
+                    // We haven't seen any messages from that validator in this era.
+                    // There can't be any in the justifications either.
+                    empty(validator).pure[F]
+                  case Some(Honest(_)) =>
+                    if (messages.nonEmpty) {
+                      import io.casperlabs.shared.Sorting.jRankOrdering
+                      // Since we know that validator is honest we can pick the newest message.
+                      honest(validator, messages.maxBy(_.jRank)).pure[F]
+                    } else {
+                      // There are no messages in the direct justifications
+                      // but we know that local DAG has seen messages from that validator.
+                      // We have to look for them in the indirect ones but it still might not be there
+                      // (because creator of the messages hasn't seen anything from that validator).
+                      DagOperations
+                        .swimlaneVFromJustifications[F](
+                          validator,
+                          validatorsLatestMessages.values.flatten.toList,
+                          dag
+                        )
+                        .takeWhile(_.eraId == era)
+                        .headOption
+                        .map(_.fold(empty(validator))(honest(validator, _)))
+                    }
+                  case Some(Equivocated(_, _)) =>
+                    if (messages.size > 1)
+                      // `messages` should be the latest messages by that validator in this era.
+                      // They are tips of his swimlane and evidences for equivocation.
+                      equivocated(validator, messages).pure[F]
+                    else {
+                      val startingPoints =
+                        validatorsLatestMessages.values.flatten.toList
+
+                      // Latest message by that validator in this era visible in the j-past-cone
+                      // of the message received.
+                      val jConeTips = DagOperations
+                        .swimlaneVFromJustifications[F](
+                          validator,
+                          startingPoints,
+                          dag
+                        )
+                        .takeWhile(_.eraId == era)
+                        .foldWhileLeft(EraObservedBehavior.unknown) {
+                          case (acc, msg) =>
+                            acc.validate(msg) match {
+                              case Undefined       => unknown.asLeft[ValidatorStatus]
+                              case h: Swimlane     => h.asLeft[ValidatorStatus]
+                              case e: Equivocation => e.asRight[ValidatorStatus]
+                            }
+                        }
+
+                      jConeTips.map {
+                        case Undefined =>
+                          // No message in the j-past-cone
+                          empty(validator)
+                        case Swimlane(tip, _) =>
+                          // ByteString is honest in the j-past-cone
+                          honest(validator, tip.m)
+                        case Equivocation(tips) =>
+                          equivocated(validator, tips.map(_.tip.m))
+                      }
+                    }
+                }
+            }
+            .map(era -> _.toMap)
+      }
+      .map(_.toMap)
+      .map { tips =>
+        val nonEmptyTips = tips.mapValues(_.filterNot(_._2.isEmpty))
+        EraObservedBehavior(nonEmptyTips).asInstanceOf[MessageJPast[Message]]
+      }
+  }
 
   /**
     * Returns latest messages per era.
